Update s6 example files
Change -t to t for s6-log scripts:
With the release of s6 v2.1.1.0 the -t flag to s6-log was depricated and
replaced with a per-match control directive. This updates the example
scripts to follow the current supported logging syntax.

Update syslogd-linux example logger:
It's helpful if you're doing syslog-style logging to split your log
streams into different files in addition to a consolidated log. This
change also splits the logging script out to be easier to manage. This
also serves a second purpose of highlighting the fact that s6-log
supports multiple log targets.

Remove (most) absolute paths for non-init replacement s6-svscan:
With non-slashpackage installs becoming the default, odds are that the
execline and s6 binaries will be installed somewhere in the path. While
all the example scripts still make calls to /command/execlineb, it makes
sense to reduce the amount of editing needed to make these scripts work.

Updating gitignore to skip trying to add vim turds