s6-softlimit: add -P option for RLIMIT_RTPRIO
Bash ulimit uses -r for this, but in s6-softlimit that was already
taken.  I could have used -R, but Bash uses that for RLIMIT_RTTIME, so
I didn't want to claim that one and create a similarly difficult
decision for somebody adding that in future.  I decided to use -P (for
"priority") as the least bad still-available option.